aggiunta sessione in sicurezza

+++ Presentation/com.uniroma2.prese.angular/src/app/app.component.ts
@@ -1,38 +1,10 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, Injector } from '@angular/core';
 import { LoaderService } from './core/services/loader.service';
 import { LoaderState } from './shared/interface/loader-state';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { SessionService } from './core/services/session.service';
+import { LoggedService } from './core/services/loggedService';
 
-@Component({
-  selector: 'ngbd-modal-content',
-  styles: ['.modal-header{background-color: red;}'] ,
-  template: `
-    <div class="modal-header">
-      <h4 class="modal-title">{{title}}</h4>
-      <button type="button" class="close" aria-label="Close" (click)="activeModal.dismiss('Cross click')">
-        <span aria-hidden="true">&times;</span>
-      </button>
-    </div>
-    <div class="modal-body">
-      <div *ngFor="let error of errors">
-        <p>{{error}}!</p>
-      </div>
-    </div>
-    <div class="modal-footer">
-      <button type="button" class="btn btn-outline-dark" (click)="activeModal.close('Close click')">Chiudi</button>
-    </div>
-  `
-})
-
-export class NgbdModalContentError {
-  @Input() errors;
-  @Input() title;
-
-  constructor(
-    public activeModal: NgbActiveModal
-  ) {}
-}
 
 @Component({
   selector: 'app-root',
@@ -46,19 +18,29 @@ export class AppComponent implements OnInit{
 
   constructor(
       private loaderService: LoaderService,
-      private sessionService: SessionService
+      private sessionService: SessionService,
+      private loggedService: LoggedService,
     ) {
+      loggedService.isLogged().subscribe(data => {
+        if(data) {
+          this.sessionService.changeToken(data);
+        }
+      })
   }
 
   ngOnInit(): void {
+
     this.loaderService.loaderState.subscribe((val: LoaderState) => {
       this.showLoader = val;
+        this.sessionService.currentSessionToken.subscribe(token => this.token = token)
     });
-    this.sessionService.currentSessionToken.subscribe(token => this.token = token)
   }
 
   logout() {
-    this.sessionService.deleteToken();
+    this.loggedService.logout().subscribe(data => {
+      if(data) {
+        this.sessionService.deleteToken();
+      }
+    })
   }
-
 }
